requirements.txt
anthropic
dbt-duckdb
langgraph
langchain-anthropic
pydantic
pandas
python-dotenv
